backend/requirements.txt
django-mongodb-backend
djangorestframework>=3.14
pymongo>=4.6
python-dotenv>=1.0.0